cdk/requirements.txt
boto3==1.24.53
aws-cdk-lib==2.93.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0.1